adding deps
